requirements.txt
requests
beautifulsoup4
python-dotenv
flask
flask_bootstrap
flask_bootstrap5
azure.cosmos
azure.storage.blob
flask_wtf
